Split util functions out for better reuse and cleaner config file
